pyproject.toml
[project]
name = "agentmemory"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langmem>=0.0.30",
    "mem0ai>=1.0.8",
    "python-dotenv>=1.2.2",
]
